projeto-holhooja/webapp_flask/requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
blinker==1.7.0
cffi==1.15.0
click==8.1.2
dnspython==2.2.1
email-validator==1.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-MQTT==1.2.1  
Flask-Session==0.4.0 
Flask-SocketIO==5.3.6 
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==21.2.0 
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
mariadb==1.1.10 
MarkupSafe==2.1.1
paho-mqtt==2.0.0 
pycparser==2.21
simple-websocket==1.0.0 
six==1.16.0
SQLAlchemy==1.4.35
sqlparse==0.4.2
Werkzeug==2.2.2
wheel==0.37.1 
WTForms==3.0.1
zipp==3.8.0
